backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
ollama==0.1.7
openai==1.3.0
python-dotenv==1.0.0
readability-lxml==0.8.1
newspaper3k==0.2.8
